🌱 Reorganize tests: separate unit and integration tests with build tags
- Add  tags to all integration tests
- Rename integration tests to *_integration_test.go convention
- Move alpha update tests from test/e2e to pkg/cli/alpha/internal/update
- Add test-integration Makefile target for integration tests
- Update test.sh to call individual test targets explicitly
- Fix test-coverage to include only plugins with tests
- Remove redundant test scripts (test/features.sh, test/integration.sh)
- Update AGENTS.md with clear test organization guidelines
