+++ Makefile
@@ -145,25 +145,21 @@ go-apidiff:
 ##@ Tests
 
 .PHONY: test
-test: test-unit test-integration test-features test-testdata test-book test-license ## Run the unit and integration tests (used in the CI)
+test: test-unit test-integration test-testdata test-book test-license ## Run the unit and integration tests (used in the CI)
 
 .PHONY: test-unit
 TEST_PKGS := ./pkg/... ./test/e2e/utils/...
 test-unit: ## Run the unit tests
 	go test -race $(TEST_PKGS)
 
-.PHONY: test-coverage
-test-coverage: ## Run unit tests creating the output to report coverage
-	- rm -rf *.out  # Remove all coverage files if exists
-	go test -race -failfast -tags=integration -coverprofile=coverage-all.out -coverpkg="./pkg/cli/...,./pkg/config/...,./pkg/internal/...,./pkg/machinery/...,./pkg/model/...,./pkg/plugin/...,./pkg/plugins/golang/...,./pkg/plugins/external/...,./pkg/plugins/optional/grafana/...,./pkg/plugins/optional/helm/v2alpha/..." $(TEST_PKGS)
-
-.PHONY: test-features
-test-features: ## Run the integration tests
-	./test/features.sh
-
 .PHONY: test-integration
-test-integration: ## Run the integration tests
-	./test/integration.sh
+test-integration: install ## Run the integration tests (requires kubebuilder binary in PATH)
+	go test -race -tags=integration -timeout 30m $(TEST_PKGS)
+
+.PHONY: test-coverage
+test-coverage: ## Run unit and integration tests with coverage report
+	- rm -rf *.out  # Remove all coverage files if exists
+	go test -race -failfast -tags=integration -timeout 30m -coverprofile=coverage-all.out -coverpkg="./pkg/cli/...,./pkg/config/...,./pkg/internal/...,./pkg/machinery/...,./pkg/model/...,./pkg/plugin/...,./pkg/plugins/golang/...,./pkg/plugins/external/...,./pkg/plugins/optional/grafana/...,./pkg/plugins/optional/helm/v2alpha/..." $(TEST_PKGS)
 
 .PHONY: check-testdata
 check-testdata: ## Run the script to ensure that the testdata is updated
