+++ pkg/plugins/optional/grafana/v1alpha/scaffolds/edit_test.go
@@ -1,3 +1,6 @@
+//go:build integration
+// +build integration
+
 /*
 Copyright 2025 The Kubernetes Authors.
