+++ pkg/plugins/optional/helm/v2alpha/scaffolds/internal/kustomize/parser_test.go
@@ -1,3 +1,6 @@
+//go:build integration
+// +build integration
+
 /*
 Copyright 2025 The Kubernetes Authors.
